bloody hell, this instructor

+++ Qsolve.h
@@ -1,4 +1,4 @@
 #include <math.h>
 #include "MyLog.h"
 
-int qsolve(double a, double b, double c, double* x1, double* x2);
+int qsolve(float a, float b, float c, float* x1, float* x2);
